fix imports and split up InfoPanel.tsx into multiple components
